+++ backend/controllers/personnagesControllers.js
@@ -20,7 +20,7 @@ export const createPersonnage = async (req, res) => {
             return res.status(400).json({ message: "La combinaison du nom et du roman doit être unique." });
 
         let newPath;
-        if (files.image && files.image.length > 0) {
+        if (files.image && files.image.length > 0 && files.image[0].originalFilename !== "perso_default.png") {
             newPath = await saveImage(files.image[0].filepath, files);
         } else {
             newPath = "images/perso_default.png";
@@ -96,7 +96,7 @@ export const updatePersonnage = async (req, res) => {
         });
 
         let newPath;
-        if (files.image && files.image.length > 0) {
+        if (files.image && files.image.length > 0 && files.image[0].originalFilename !== "perso_default.png") {
             newPath = await saveImage(files.image[0].filepath, files);           
             await deleteImage(personnageModel, id);
         } else {
